requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
cfgv==3.5.0
distlib==0.4.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
filelock==3.24.3
identify==2.6.16
nodeenv==1.10.0
platformdirs==4.9.2
pre_commit==4.5.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
python-ulid==3.1.0
PyYAML==6.0.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ulid-py==1.1.0
virtualenv==20.38.0
